Move `include/channel/ProtocolHelper.hpp` to `include/elrond_protocol_helper.hpp`
Issue #38
